User is now authenticated
Data is pulled at the build of the login page. It is stored in the User class in the User.dart file under a corresponding list. The email/password lists are checked on the login button press. If it is valid, then it moves on to the corresponding course landing page.